requirements.txt
pip==23.3.1
gunicorn==21.2.0
termcolor==2.3.0
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.44.3
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
h5py==3.10.0
itsdangerous==2.1.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
matplotlib==3.8.1
ml-dtypes==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
Pillow==10.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.15.1
tensorboard-data-server==0.7.2

